Update the get cart item function
Hi, Nathan! For this function, what I am thinking is using find user function in the database to find the user. Then using gameRes.cart to get the cart arrray of the account. But the problem is keeping the same, the cart page only re-render with correct info after the gamer log out and login again. I will join John's office hout this friday with this problem, and try to find a solution. At this time, we can display this feature in the video as gamer need to log out and login again to see the cart after adding or deleting, which is obviously a silly problem in my part. But I am okay with this at this point, I will definitely improve this for the final version.